requirements.txt
adafruit-circuitpython-gps
pyserial
pygame
